+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Remove support for building winesapOS images based on SteamOS 3.
 - Change Pacman to not ignore any package updates by default.
 - Change 'pacman-static' to be installed as a package instead of a single binary.
+- Add 'curl-static' binary.
 
 ## 3.4.0 - The Game Mode Update